Adds HTTP Strict Transport Security as express middleware. This will ensure the clients using CDAP UI will use TLS enabled UI for the max-age we specified. As soon as the browser sees the hsts in the header it will make sure the TLS enabled CDAP UI gets used until max age specified.
This is a safe change when CDAP UI by default gets served in non-TLS enabled environment (meaning only http)

This is however not safe when clients change from https to http. Then the browser will try to go to https until the
max-age and will probably error out. Since downgrading to non-TLS enabled CDAP UI is not standard we are making this
change for now.